requirements.txt
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
Flask==1.1.2
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
swagger-spec-validator==2.7.3
urllib3==1.26.5
Werkzeug==1.0.1
pyjwt==2.1.0
cryptography==3.3.2
Flask-KVSession==0.6.2
webargs==6.1.1
oauthlib==3.1.0
requests_oauthlib
swagger-ui-bundle # remove for production
flask_sqlalchemy
sqlalchemy
psycopg2-binary
python-jose
pymongo==3.10.1
mongoengine==0.20.0
passlib==1.7.1
